thirteenth commit: manga by tags, token handler fix

+++ composeApp/src/commonMain/kotlin/api/mangadex/service/TokenHandler.kt
@@ -8,17 +8,17 @@ import kotlin.io.encoding.ExperimentalEncodingApi
 interface TokenHandler {
     companion object {
         @OptIn(ExperimentalEncodingApi::class)
-        fun expiration(token: Token): Int {
+        fun expiration(token: Token): Long {
             val splitted = token.accessToken.split(".")
             val payload = Base64.decode(splitted[1]).decodeToString()
-            return payload.substringBefore(',').substringAfter(':').toInt()
+            return payload.substringBefore(',').substringAfter(':').toLong()
         }
 
         @OptIn(ExperimentalEncodingApi::class)
-        fun expiration(token: String): Int {
+        fun expiration(token: String): Long {
             val splitted = token.split(".")
             val payload = Base64.decode(splitted[1]).decodeToString()
-            return payload.substringBefore(',').substringAfter(':').toInt()
+            return payload.substringBefore(',').substringAfter(':').toLong()
         }
     }
